+++ src/test/java/roomescape/domain/theme/ThemeAggregationLimitTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.theme;
 
 import static org.assertj.core.api.Assertions.assertThat;
